Split stuff to delay import
